+++ libs/asset_type_lightsail.py
@@ -37,7 +37,7 @@ class LIGHTSAIL(AssetType):
         self.role_admin = ''
         self.instance_id = ''
 
-    def report(self, report, brief=False):
+    def report(self, report, brief=False, with_fpkey=False):
         """
         Add an asset with only relevent informations
         """
@@ -64,7 +64,7 @@ class LIGHTSAIL(AssetType):
             if self.role_admin:
                 asset_report['Roles Admin'] = f'[red]{self.role_admin}[/red]'
             if self.security_issues:
-                self.update_audit_report(asset_report)
+                self.update_audit_report(asset_report, with_fpkey)
         if 'LIGHTSAIL' not in report[self.location.region][self.location.vpc][self.location.subnet]:
             report[self.location.region][self.location.vpc][self.location.subnet]['LIGHTSAIL'] = \
                 { self.name: asset_report }