+++ libs/asset_type_rds.py
@@ -25,7 +25,7 @@ class RDS(AssetType):
         self.engine = engine
         self.url = url
 
-    def report(self, report, brief=False):
+    def report(self, report, brief=False, with_fpkey=False):
         """
         Add an asset with only relevent informations
         """
@@ -40,7 +40,7 @@ class RDS(AssetType):
             if self.url:
                 asset_report['URL'] = self.url
             if self.security_issues:
-                self.update_audit_report(asset_report)
+                self.update_audit_report(asset_report, with_fpkey)
         if 'RDS' not in report[self.location.region][self.location.vpc][self.location.subnet]:
             report[self.location.region][self.location.vpc][self.location.subnet]['RDS'] = \
                 { self.name: asset_report }