+++ libs/tools.py
@@ -8,6 +8,7 @@ Written by Nicolas BEGUIER (nicolas.beguier@adevinta.com)
 """
 
 # Standard library imports
+import hashlib
 import json
 import logging
 from pathlib import Path
@@ -166,6 +167,19 @@ def get_account_in_arn(arn):
     if len(arn.split(':')) >= 5:
         return arn.split(':')[4]
     return '000000000000'
+
+def get_false_positive_key(message, asset_type, asset_name):
+    """
+    Returns the type, name and the 10 first characters of the message
+    """
+    # Convert the input string to bytes, as hashlib works with bytes
+    input_bytes = message.encode('utf-8')
+
+    # Compute the SHA-1 hash of the input bytes
+    sha1_hash = hashlib.sha1(input_bytes).hexdigest()
+
+    # Return the first 10 characters of the SHA-1 hash
+    return f'{asset_type.lower()}-{asset_name.lower().replace(" ", "-")}-{sha1_hash[:10]}'
 
 def get_lambda_name(apigw_arn):
     """