+++ libs/asset_type_cf.py
@@ -57,7 +57,7 @@ class CloudFront(AssetType):
             self.authorization.add_auth_type(auth_type)
         self.location.region = 'global'
 
-    def report(self, report, brief=False):
+    def report(self, report, brief=False, with_fpkey=False):
         """
         Add an asset with only relevent informations
         """
@@ -71,7 +71,7 @@ class CloudFront(AssetType):
             if self.public:
                 asset_report['PubliclyAccessible'] = '[red]True[/red]'
             if self.security_issues:
-                self.update_audit_report(asset_report)
+                self.update_audit_report(asset_report, with_fpkey)
         if 'CloudFront' not in report[self.location.region]:
             report[self.location.region]['CloudFront'] = { self.name: asset_report }
             return report