+++ libs/asset_type_ec2.py
@@ -38,7 +38,7 @@ class EC2(AssetType):
         self.instance_id = ''
         self.eks_cluster = ''
 
-    def report(self, report, brief=False):
+    def report(self, report, brief=False, with_fpkey=False):
         """
         Add an asset with only relevent informations
         """
@@ -65,13 +65,16 @@ class EC2(AssetType):
             if self.role_admin:
                 asset_report['Roles Admin'] = f'[red]{self.role_admin}[/red]'
             if self.security_issues:
-                self.update_audit_report(asset_report)
+                self.update_audit_report(asset_report, with_fpkey)
         if 'EC2' not in report[self.location.region][self.location.vpc][self.location.subnet]:
             report[self.location.region][self.location.vpc][self.location.subnet]['EC2'] = \
                 { self.name: asset_report }
             return report
+        reported_name = self.name
+        if self.name in report[self.location.region][self.location.vpc][self.location.subnet]['EC2']:
+            reported_name = f'{self.name}_{self.instance_id}'
         report[self.location.region][self.location.vpc][self.location.subnet]['EC2'].update(
-            { self.name: asset_report })
+            { reported_name: asset_report })
         return report
 
     def report_brief(self):