+++ libs/display.py
@@ -49,7 +49,7 @@ def prepare_report(assets, meta_types, console):
     return report
 
 @log_me('Auditing the scan...')
-def audit_scan(assets, report, security_config, brief, console):
+def audit_scan(assets, report, security_config, brief, with_fpkey, console):
     """
     Add assets in report and audit them
     """
@@ -70,17 +70,20 @@ def audit_scan(assets, report, security_config, brief, console):
             if not asset.security_issues:
                 continue
             asset.remove_not_vulnerable_members()
-        report = asset.report(report, brief=brief)
+            asset.remove_false_positives()
+        if security_config and not asset.security_issues:
+            continue
+        report = asset.report(report, brief=brief, with_fpkey=with_fpkey)
     return report
 
-def print_report(assets, meta_types, console, output_file, brief=False, security_config=None):
+def print_report(assets, meta_types, console, output_file, brief=False, with_fpkey=False, security_config=None):
     """
     Print subnets
     """
     # Construct Region/VPC/subnet
     report = prepare_report(assets, meta_types, console)
 
-    report = audit_scan(assets, report, security_config, brief, console)
+    report = audit_scan(assets, report, security_config, brief, with_fpkey, console)
 
     str_report = json.dumps(report, sort_keys=True, indent=4)
 
@@ -151,7 +154,7 @@ def draw_vpc_peering(assets, dot_filename, args):
 def draw_threats(title, assets, csl, args):
     # Diagrams imports
     from diagrams import Diagram, Cluster
-    from diagrams.aws.compute import EC2, EKS, LambdaFunction as Lambda
+    from diagrams.aws.compute import EC2, EKS, LambdaFunction as Lambda, Lightsail as LIGHTSAIL
     from diagrams.aws.network import ELB, CloudFront, APIGateway as APIGW, VPC
     from diagrams.aws.database import RDS
     from diagrams.aws.storage import S3