+++ libs/scan.py
@@ -17,6 +17,7 @@ import libs.asset_type_ec2 as ec2
 import libs.asset_type_eks as eks
 import libs.asset_type_elb as elb
 import libs.asset_type_iam_group as iam
+import libs.asset_type_lightsail as lightsail
 import libs.asset_type_mq as mq
 import libs.asset_type_rds as rds
 import libs.asset_type_route53 as r53
@@ -39,6 +40,7 @@ def get_raw_data(boto_session, meta_types, cache, console):
         'eks': True,
         'elb': True,
         'iam': True,
+        'lightsail': True,
         'mq': True,
         'rds': True,
         's3': True,
@@ -80,6 +82,14 @@ def get_raw_data(boto_session, meta_types, cache, console):
 
     if 'ELB' in meta_types:
         raw_data, authorizations = elb.get_raw_data(
+            raw_data,
+            authorizations,
+            boto_session,
+            cache,
+            console)
+
+    if 'LIGHTSAIL' in meta_types:
+        raw_data, authorizations = lightsail.get_raw_data(
             raw_data,
             authorizations,
             boto_session,
@@ -213,6 +223,16 @@ def aws_scan(
             iam_action_passlist,
             iam_rolename_passlist,
             name_filter,
+            cache,
+            console)
+
+    if 'LIGHTSAIL' in meta_types:
+        assets, authorizations = lightsail.parse_raw_data(
+            assets,
+            authorizations,
+            raw_data,
+            name_filter,
+            public_only,
             cache,
             console)