+++ CHANGELOG.md
@@ -14,6 +14,9 @@ AWS-TOWER
   - Add `--false-positive-key` in audit mode
   - Add `config/false_positives_list.txt` file, list of finding keys to ignore in audit report
 
+### Fixtures
+  - Audit: Consider ports 0-65535 as "all"
+
 4.4.5
 -----
 2023/07/28